#55 Refatorações removendo as actions do SistemaController e distribuindo para outros Controller. Refatorações dos métodos de CompraServico relacionados à compra de ingressos.

+++ src/main/java/br/ufg/inf/pitanga/dao/CompraDao.java
@@ -41,7 +41,7 @@ public class CompraDao implements InterfaceCompraDao {
     }
 
     @Override
-    public Compra buscarPorId(int id) {
+    public Compra buscarPorId(Long id) {
         return manager.find(Compra.class, id);
     }
